Power6Backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.23.2
httpx==0.27.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4


# Optional tools & CLI
click==8.1.7
typer==0.9.0
colorama==0.4.6
packaging==23.2

# Data/Parsing
pyparsing==3.1.2
pyyaml==6.0.1

# Background / scheduling
anyio==4.3.0

# SQLite + Testing
sqlmodel==0.0.16
pytest==8.2.2
